Command line option argument for i'm feeling lucky mode
