+++ src/main.c
@@ -4,157 +4,11 @@
 #include <stdlib.h> 
 #include <string.h> 
 #include <stdbool.h>
+#include <unistd.h>
 #include "../include/parser.h"
+#include "../include/communication.h"
 #include "../include/util.h"
-
-#define PAGE_SIZE 4096 
-#define REQ_SIZE 256
 #define DEBUG 1
-#define GET_REDIRECTS 0x00
-#define GET_CONTENT 0x01
-#define GET_DISAMBIGUATION 0x02
-#define GET_SEARCH 0x04
-#define RESP_OK 0xFF
-
-typedef struct reqdata {
-
-  //Lets the underlying function know what API request to make
-  int req_type;
-
-  //Lets the user know what to do next
-  int res_type;
-
-  //Response makes it clear that this is a disambiguation
-  char * next_choice; 
-
-} reqdata; 
-
-
-/**
- * @brief Handle data from the web server.
- * @param ptr string of raw data received
- * @param size number of pages to parse
- * @param size of the page
- * @param userdata pointer to data provided by the user
- */
-static size_t handle_resp (char *ptr, size_t size, size_t nmemb, void * userdata)
-{
-  if ((NULL == ptr) || (NULL == userdata))
-  {
-    return 0; 
-  }
-
-  if (GET_CONTENT == ((reqdata *) userdata)->req_type)
-  {
-    char * content = parse_content_from_json(ptr); 
-
-    if (NULL != content)
-    {
-      if (NULL != strstr(content, "may refer to"))
-      {
-        ((reqdata *) userdata)->res_type = GET_DISAMBIGUATION;
-        printf("There are multiple pages.  Try one of these:\n\n"); 
-        return 0;
-      }
-      else
-      {
-        ((reqdata *) userdata)->res_type = RESP_OK;
-        printf("%s\n",content); 
-      }
-    }
-  }
-  else if (GET_SEARCH == ((reqdata *) userdata)->req_type)
-  {
-    char * req = parse_search(ptr); 
-    ((reqdata *) userdata)->next_choice = req;
-  }
-
-  //Disambiguation
-  else
-  {
-    char * req = parse_disambiguation(ptr); 
-    ((reqdata *) userdata)->next_choice = req;
-  }
-
-  return (size * nmemb);  
-}
-
-
-/**
- * @brief function to actually request the given page
- * @param wiki_page wiki page string to request
- * @param request_data data provided by the requester to this function
- */
-void make_request(char * wiki_page, reqdata * request_data)
-{
-  char * request_url = calloc(REQ_SIZE, sizeof(char)); 
-  CURL * curl;
-  curl = curl_easy_init();
-
-  if ((NULL == curl) || (NULL == request_url)) 
-  {
-    return;
-  }
-
-  char * base_wiki_url = "https://en.wikipedia.org/w/api.php";
-  char * com;
-  char * flags = "&formatversion=2&format=json"; 
-
-  //Content query command
-  if (GET_CONTENT == request_data->req_type)
-  {
-    com = "?action=query&prop=extracts&exsentences=10&exlimit=5&titles=";
-  }
-
-  //Search query command
-  else if (GET_SEARCH == request_data->req_type)
-  {
-    com = "?action=query&list=search&srsearch=";
-  }
-
-  //Disambiguation query command
-  else
-  {
-    printf("Disambiguation...\n"); 
-    com = "?action=query&prop=links&pllimit=10&titles=";
-  }
-
-  strncpy(request_url, base_wiki_url, strlen(base_wiki_url)); 
-  strncat(request_url, com, strlen(com)); 
-
-  //Strip quotes from wiki_page: 
-  wiki_page = remove_char(wiki_page, '"'); 
-  wiki_page = replace_spaces(wiki_page); 
-  strncat(request_url, wiki_page, strlen(wiki_page)); 
-  strncat(request_url, flags, strlen(flags));
-
-  //printf("Request URL: %s\n",request_url); 
-  curl_easy_setopt(curl, CURLOPT_URL, request_url);
-
-  //Follow redirects
-  curl_easy_setopt(curl, CURLOPT_FOLLOWLOCATION, 1L);
-
-  //Data to pass to the handler function
-  curl_easy_setopt(curl, CURLOPT_WRITEDATA, (void *) request_data);
-
-#ifdef DEBUG
-  //printf("Request URL: %s\n",request_url); 
-#endif
-
-  //Function for handling response
-  curl_easy_setopt(curl, CURLOPT_WRITEFUNCTION, handle_resp);
-
-  //Perform the actual request
-  CURLcode res = curl_easy_perform(curl);
-
-  if (CURLE_OK != res)
-  {
-    fprintf(stderr, "err: %s\n", curl_easy_strerror(res));
-  }
-
-  curl_easy_cleanup(curl);
-  free(request_url); 
-}
 
 int main(int argc, char ** argv)
 {
@@ -167,14 +21,45 @@ int main(int argc, char ** argv)
     usage(); 
     return 1; 
   }
+  
+  bool lucky = false;
+  int opt;
+
+  while (-1 != (opt = getopt(argc, argv, "l")))
+  {
+
+    switch (opt)
+    {
+      case 'l':
+        lucky = true;
+      break;
+
+      default: 
+        usage();
+        return 1;
+      break;
+    }
+  }
 
   //Get the desired page from the user: 
-  char * wiki_page = combine_args_to_page(argc, argv);
+  //optind is the index given after regular args
+  printf("Optind: %d\n",optind);
+
+  if (optind >= argc)
+  {
+    usage();
+    return 1;
+  }
+
+  char * wiki_page;
+  wiki_page = combine_args_to_page(argc - (optind - 1), argv + (optind - 1));
 
   //First, search for the page
   reqdata request;
   request.next_choice = NULL;
   request.req_type = GET_SEARCH;
+  request.res_type = RESP_OK;
+  request.is_lucky = lucky;
 
   make_request(wiki_page, &request); 
   
@@ -187,8 +72,11 @@ int main(int argc, char ** argv)
       make_request(wiki_page, &request); 
     }
 
-    //If there
-    if (NULL != request.next_choice)
+    if (NULL == request.next_choice)
+    {
+      break;
+    }
+    else
     {
       request.req_type = GET_CONTENT;
       make_request(capitalize_arg(request.next_choice), &request); 
@@ -196,6 +84,7 @@ int main(int argc, char ** argv)
   }
   
   free(wiki_page); 
+
   return 0;
 }
